refactor(header-chain): split engine, invariants, and recovery audit by responsibility
Break three large transition modules into facade packages with private
children so each file owns one concern, without changing the public
surface re-exported from transition/mod.rs.

- engine/: HeaderChainEngine + plan/install in mod.rs; TransitionInput
  facts in input.rs; projection/aux install helpers (+ tests) in install.rs
- invariants/: verify_candidate / mode dispatch in mod.rs; incremental
  aux-auth and checkpoint-finality fast paths colocated; leaf checks under
  checks/{nodes,indexes,projections,pins,protected,generations,aux}.rs
- recovery/audit/: audit_authoritative facade with nodes, connectivity,
  trust_pins, and authoritative_rows clusters (kept separate from commit
  invariants)
- rename recovery/model.rs → phases.rs for phase carriers only
