--- crates/zakura-header-chain/src/transition/invariants.rs
@@ -1,998 +1,0 @@
-//! Bounded commit-time verification of every projected transition invariant.
-
-use std::collections::{HashMap, HashSet};
-
-use thiserror::Error;
-use zakura_chain::block;
-
-use crate::graph::GraphOverlay;
-use crate::graph::HeaderGraphView;
-#[cfg(test)]
-use crate::TransitionPlan;
-use crate::{
-    AuxDelta, BodyValidationState, EligibilityReason, EngineMode, FinalitySource, Frontier,
-    HeaderChainEngine, HeaderNode, PlanCandidate, ProjectionDelta,
-};
-
-/// Stable, category-specific projected-state invariant failures.
-#[derive(Copy, Clone, Debug, Eq, Error, PartialEq)]
-pub enum InvariantViolation {
-    /// 1. The verifier found conflicting row-key, canonical-header, and computed hashes.
-    #[error("node hash invariant failed at {0:?}")]
-    NodeHash(block::Hash),
-    /// 2. The verifier found a non-anchor node without an exact height-minus-one parent.
-    #[error("parent invariant failed at {0:?}")]
-    Parent(block::Hash),
-    /// 3. The verifier could not round-trip a hash, parent/child, height, or planned index.
-    #[error("index invariant failed at {0:?}")]
-    Index(block::Hash),
-    /// 4. The verifier found an incorrect work origin or parent-plus-block value.
-    #[error("work invariant failed at {0:?}")]
-    Work(block::Hash),
-    /// 5. The verifier found cached inherited eligibility that differs from exact ancestry.
-    #[error("eligibility invariant failed at {0:?}")]
-    Eligibility(block::Hash),
-    /// 6. The verifier found a gap in the finalized-to-tip selected projection.
-    #[error("selected projection invariant failed at {0:?}")]
-    SelectedProjection(block::Hash),
-    /// 7. The verifier found an eligible score above `header_best`.
-    #[error("selection invariant failed")]
-    Selection,
-    /// 8. The verifier found a verified projection that contradicts its mode or body evidence.
-    #[error("verified projection invariant failed at {0:?}")]
-    VerifiedProjection(block::Hash),
-    /// 9. The verifier found a retained path that conflicts with an authenticated trust pin.
-    #[error("trust-pin invariant failed at height {0:?}")]
-    TrustPin(block::Height),
-    /// 10. The transition evicted finalized, selected, or verified protected state.
-    #[error("protected-path invariant failed at {0:?}")]
-    Protected(block::Hash),
-    /// 11. The projected DAG exceeds a frozen resource limit.
-    #[error("resource-limit invariant failed")]
-    Limits,
-    /// 12. The verifier found generation increments that disagree with actual changes.
-    #[error("generation invariant failed")]
-    Generation,
-    /// 13. The verifier found auxiliary evidence without a retained foreign key or provenance link.
-    #[error("auxiliary invariant failed at {0:?}")]
-    Auxiliary(block::Hash),
-    /// The coherent snapshot before commit changed or failed during plan verification.
-    #[error("snapshot before commit changed during invariant verification")]
-    SnapshotBeforeCommit,
-}
-
-/// Selects which projected graph and node set the verifier inspects.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-enum VerificationMode {
-    /// Materialized projected graph and every retained node (test/fuzz oracle).
-    #[cfg(any(test, feature = "fuzz-impl"))]
-    Exhaustive,
-    /// Delta overlay and changed-boundary nodes (shipped production path).
-    #[cfg(any(test, not(feature = "fuzz-impl")))]
-    Production,
-}
-
-fn default_verification_mode() -> VerificationMode {
-    #[cfg(any(test, feature = "fuzz-impl"))]
-    {
-        VerificationMode::Exhaustive
-    }
-    #[cfg(not(any(test, feature = "fuzz-impl")))]
-    {
-        VerificationMode::Production
-    }
-}
-
-/// Independently check that `plan`'s projection obeys every transition invariant under `engine_before_commit`.
-///
-/// Pure gate between [`PlanCandidate`] and [`TransitionPlan`]: no mutation; success is required
-/// before `TransitionPlan::from_verified`; failure is [`InvariantViolation`].
-pub(crate) fn verify_candidate(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-) -> Result<(), InvariantViolation> {
-    verify_plan_with_mode(engine_before_commit, plan, default_verification_mode())
-}
-
-fn verify_plan_with_mode(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    if is_incremental_aux_authentication(engine_before_commit, plan) {
-        return verify_incremental_aux_authentication(engine_before_commit, plan, mode);
-    }
-    if is_incremental_checkpoint_finality(engine_before_commit, plan) {
-        return verify_incremental_checkpoint_finality(engine_before_commit, plan, mode);
-    }
-
-    verify_plan_exhaustive(engine_before_commit, plan, mode)
-}
-
-fn verify_plan_exhaustive(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    let source = engine_before_commit.snapshot();
-    if source != plan.snapshot_before_commit {
-        return Err(InvariantViolation::SnapshotBeforeCommit);
-    }
-    engine_before_commit
-        .graph()
-        .validate_delta(plan.graph_delta())
-        .map_err(|_| InvariantViolation::Index(block::Hash([0; 32])))?;
-    let source_metadata = engine_before_commit.metadata();
-    let delta_graph = GraphOverlay::from_delta(engine_before_commit.graph(), plan.graph_delta())
-        .map_err(|_| InvariantViolation::Index(block::Hash([0; 32])))?;
-    let delta_finalized = delta_graph.view_finalized_frontier();
-    #[cfg(any(test, feature = "fuzz-impl"))]
-    if mode == VerificationMode::Exhaustive {
-        let projected_graph = materialize_projected_graph(engine_before_commit, plan)?;
-        return verify_plan_against_graph(
-            engine_before_commit,
-            plan,
-            &source,
-            source_metadata,
-            &projected_graph,
-            delta_finalized,
-            mode,
-        );
-    }
-    verify_plan_against_graph(
-        engine_before_commit,
-        plan,
-        &source,
-        source_metadata,
-        &delta_graph,
-        delta_finalized,
-        mode,
-    )
-}
-
-fn verify_plan_against_graph<G: HeaderGraphView>(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    source: &crate::EngineSnapshot,
-    source_metadata: &crate::EngineMetadata,
-    graph: &G,
-    delta_finalized: Frontier,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    let metadata = &plan.change_set.metadata;
-    let expected_work_origin = if plan.graph_delta().rebases_work_coordinates() {
-        source.frontiers.finalized
-    } else {
-        source_metadata.work_origin
-    };
-    if source_metadata.state_version != source.state_version
-        || metadata.mode != source.mode
-        || metadata.work_origin != expected_work_origin
-    {
-        return Err(InvariantViolation::SnapshotBeforeCommit);
-    }
-    if metadata.frontiers.finalized != graph.view_finalized_frontier()
-        || metadata.frontiers.finalized != delta_finalized
-        || metadata.frontiers.finalized.height < source.frontiers.finalized.height
-        || match plan.change_set.finality_append {
-            Some(record) => {
-                record.previous != source.frontiers.finalized
-                    || record.current != metadata.frontiers.finalized
-                    || record.epoch != metadata.finality_epoch
-            }
-            None => metadata.frontiers.finalized != source.frontiers.finalized,
-        }
-    {
-        return Err(InvariantViolation::Protected(
-            metadata.frontiers.finalized.hash,
-        ));
-    }
-    if let Some(record) = plan.change_set.finality_append {
-        let valid_source = match record.source {
-            FinalitySource::FullState { .. } => metadata.mode == EngineMode::Integrated,
-            FinalitySource::HeadersOnlyDepth { selected_tip } => {
-                metadata.mode == EngineMode::HeadersOnly
-                    && selected_tip
-                        .height
-                        .0
-                        .saturating_sub(record.current.height.0)
-                        == plan.limits.local_finality_depth.get()
-                    && graph
-                        .view_header_ancestor(selected_tip.hash, record.current.height)
-                        .ok()
-                        .flatten()
-                        == Some(record.current)
-            }
-            FinalitySource::MigratedHeadersOnly => true,
-        };
-        if !valid_source {
-            return Err(InvariantViolation::Protected(record.current.hash));
-        }
-    } else if metadata.finality_epoch != source_metadata.finality_epoch {
-        return Err(InvariantViolation::Generation);
-    }
-    for node in verification_nodes(engine_before_commit, graph, plan, mode) {
-        verify_node(graph, node, metadata.work_origin.hash)?;
-    }
-    verify_indexes(engine_before_commit, plan)?;
-    let selected = projected_path(engine_before_commit, source, &plan.change_set.selected_projection, true)?;
-    let verified = projected_path(engine_before_commit, source, &plan.change_set.verified_projection, false)?;
-    verify_projection(
-        graph,
-        &selected,
-        metadata.frontiers.header_best,
-        InvariantViolation::SelectedProjection,
-    )?;
-    let best = graph
-        .view_select_best_header_chain()
-        .map_err(|_| InvariantViolation::Selection)?;
-    if best.0 != metadata.frontiers.header_best || best.1 != metadata.header_best_score {
-        return Err(InvariantViolation::Selection);
-    }
-    verify_verified(
-        graph,
-        metadata.mode,
-        &verified,
-        metadata.frontiers.verified_best,
-    )?;
-    verify_pins(
-        &plan.trust_pins,
-        &selected,
-        &verified,
-        &plan.change_set.put_nodes,
-    )?;
-    verify_protected(graph, plan)?;
-    if graph.view_header_node_count().saturating_sub(1) > plan.limits.max_non_finalized_nodes.get()
-        || graph.view_eligible_header_tips().len() > plan.limits.max_candidate_tips.get()
-    {
-        return Err(InvariantViolation::Limits);
-    }
-    verify_generations(engine_before_commit, plan, &selected, &verified)?;
-    verify_aux(engine_before_commit, graph, plan, mode)?;
-    Ok(())
-}
-
-pub(super) fn is_incremental_aux_authentication(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-) -> bool {
-    let metadata = &plan.change_set.metadata;
-    let source_metadata = engine_before_commit.metadata();
-
-    plan.effect().is_aux_authentication()
-        && !plan.change_set.aux_changes.is_empty()
-        && plan.change_set.aux_changes.len() <= 2
-        && plan
-            .change_set
-            .aux_changes
-            .iter()
-            .all(|change| matches!(change, AuxDelta::Put(_)))
-        && plan.change_set.put_nodes.is_empty()
-        && plan.change_set.delete_nodes.is_empty()
-        && plan.change_set.index_changes.inserted.is_empty()
-        && plan.change_set.index_changes.deleted.is_empty()
-        && plan.change_set.selected_projection == ProjectionDelta::default()
-        && plan.change_set.verified_projection == ProjectionDelta::default()
-        && plan.change_set.eligibility_changes.is_empty()
-        && plan.change_set.finality_append.is_none()
-        && plan.graph_delta().is_empty()
-        && metadata.disk_format == source_metadata.disk_format
-        && metadata.mode == source_metadata.mode
-        && metadata.network_id == source_metadata.network_id
-        && metadata.anchor_manifest_digest == source_metadata.anchor_manifest_digest
-        && metadata.work_origin == source_metadata.work_origin
-        && metadata.finality_epoch == source_metadata.finality_epoch
-        && metadata.frontiers == source_metadata.frontiers
-        && metadata.header_best_score == source_metadata.header_best_score
-        && metadata.oldest_retained_height == source_metadata.oldest_retained_height
-        && metadata.alarms == source_metadata.alarms
-}
-
-fn verify_incremental_aux_authentication(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    if engine_before_commit.snapshot() != plan.snapshot_before_commit {
-        return Err(InvariantViolation::SnapshotBeforeCommit);
-    }
-    for change in &plan.change_set.aux_changes {
-        let AuxDelta::Put(delivery) = change else {
-            return Err(InvariantViolation::Auxiliary(block::Hash([0; 32])));
-        };
-        let existing = engine_before_commit
-            .aux_deliveries(delivery.header_hash)
-            .iter()
-            .find(|existing| existing.delivery_id == delivery.delivery_id)
-            .ok_or(InvariantViolation::Auxiliary(delivery.header_hash))?;
-        let mut expected = *existing;
-        expected.authentication = delivery.authentication;
-        if expected != **delivery
-            || existing.authentication != crate::AuxAuthentication::Unauthenticated
-            || delivery.authentication == crate::AuxAuthentication::Unauthenticated
-        {
-            return Err(InvariantViolation::Auxiliary(delivery.header_hash));
-        }
-    }
-    verify_generations(
-        engine_before_commit,
-        plan,
-        engine_before_commit.selected_projection(),
-        engine_before_commit.verified_projection(),
-    )?;
-    verify_aux(engine_before_commit, engine_before_commit.graph(), plan, mode)?;
-    Ok(())
-}
-
-pub(super) fn is_incremental_checkpoint_finality(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-) -> bool {
-    let source = engine_before_commit.snapshot();
-    let metadata = &plan.change_set.metadata;
-    let Some(record) = plan.change_set.finality_append else {
-        return false;
-    };
-    let finalized = record.current;
-
-    plan.effect().is_checkpoint_finality()
-        && source.mode == EngineMode::Integrated
-        && matches!(record.source, FinalitySource::FullState { .. })
-        && record.previous == source.frontiers.finalized
-        && finalized.height > source.frontiers.finalized.height
-        && metadata.frontiers.finalized == finalized
-        && metadata.frontiers.verified_best == finalized
-        && metadata.frontiers.header_best == source.frontiers.header_best
-        && engine_before_commit.verified_projection() == [source.frontiers.finalized]
-        && engine_before_commit
-            .selected_projection()
-            .binary_search_by_key(&finalized.height, |frontier| frontier.height)
-            .ok()
-            .is_some_and(|index| engine_before_commit.selected_projection()[index] == finalized)
-        && plan.change_set.selected_projection
-            == (ProjectionDelta {
-                remove_before: Some(finalized.height),
-                remove_from: None,
-                put: Vec::new(),
-            })
-        && plan.change_set.verified_projection
-            == (ProjectionDelta {
-                remove_before: Some(finalized.height),
-                remove_from: Some(finalized.height),
-                put: vec![finalized],
-            })
-        && plan.change_set.put_nodes.len() == 1
-        && plan.change_set.put_nodes[0].hash == finalized.hash
-        && plan.change_set.eligibility_changes.is_empty()
-        && plan
-            .change_set
-            .aux_changes
-            .iter()
-            .all(|change| matches!(change, AuxDelta::Delete { .. }))
-        && plan.graph_delta().finalized_frontier() == Some(finalized)
-}
-
-fn verify_incremental_checkpoint_finality(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    let source = engine_before_commit.snapshot();
-    if source != plan.snapshot_before_commit {
-        return Err(InvariantViolation::SnapshotBeforeCommit);
-    }
-    engine_before_commit
-        .graph()
-        .validate_delta(plan.graph_delta())
-        .map_err(|_| InvariantViolation::Index(block::Hash([0; 32])))?;
-    let delta_graph = GraphOverlay::from_delta(engine_before_commit.graph(), plan.graph_delta())
-        .map_err(|_| InvariantViolation::Index(block::Hash([0; 32])))?;
-    let delta_finalized = delta_graph.view_finalized_frontier();
-    #[cfg(any(test, feature = "fuzz-impl"))]
-    if mode == VerificationMode::Exhaustive {
-        let projected_graph = materialize_projected_graph(engine_before_commit, plan)?;
-        return verify_incremental_checkpoint_against_graph(
-            engine_before_commit,
-            plan,
-            &source,
-            &projected_graph,
-            delta_finalized,
-            mode,
-        );
-    }
-    verify_incremental_checkpoint_against_graph(
-        engine_before_commit,
-        plan,
-        &source,
-        &delta_graph,
-        delta_finalized,
-        mode,
-    )
-}
-
-fn verify_incremental_checkpoint_against_graph<G: HeaderGraphView>(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    source: &crate::EngineSnapshot,
-    graph: &G,
-    delta_finalized: Frontier,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    let metadata = &plan.change_set.metadata;
-    let record = plan
-        .change_set
-        .finality_append
-        .expect("the incremental checkpoint shape requires finality evidence");
-    let finalized = record.current;
-    let source_metadata = engine_before_commit.metadata();
-
-    if source_metadata.state_version != source.state_version
-        || metadata.mode != source.mode
-        || metadata.work_origin != source_metadata.work_origin
-    {
-        return Err(InvariantViolation::SnapshotBeforeCommit);
-    }
-    if metadata.frontiers.finalized != graph.view_finalized_frontier()
-        || metadata.frontiers.finalized != delta_finalized
-        || record.previous != source.frontiers.finalized
-        || record.current != metadata.frontiers.finalized
-        || record.epoch != metadata.finality_epoch
-        || !matches!(record.source, FinalitySource::FullState { .. })
-    {
-        return Err(InvariantViolation::Protected(finalized.hash));
-    }
-
-    let changed = &plan.change_set.put_nodes[0];
-    let previous = engine_before_commit
-        .graph()
-        .header_node(changed.hash)
-        .ok_or(InvariantViolation::Protected(changed.hash))?;
-    if changed.hash != previous.hash || changed.header != previous.header {
-        return Err(InvariantViolation::NodeHash(changed.hash));
-    }
-    if changed.height != previous.height || changed.parent_hash != previous.parent_hash {
-        return Err(InvariantViolation::Parent(changed.hash));
-    }
-    if changed.block_work != previous.block_work
-        || changed.work_coordinate() != previous.work_coordinate()
-    {
-        return Err(InvariantViolation::Work(changed.hash));
-    }
-    if changed.validation != previous.validation
-        || changed.eligibility.direct_reasons != previous.eligibility.direct_reasons
-        || changed.eligibility.inherited_from.is_some()
-    {
-        return Err(InvariantViolation::Eligibility(changed.hash));
-    }
-    if changed.aux_delivery_ids != previous.aux_delivery_ids {
-        return Err(InvariantViolation::Auxiliary(changed.hash));
-    }
-    if !matches!(
-        changed.body_validation_state,
-        BodyValidationState::Verified { .. }
-    ) {
-        return Err(InvariantViolation::VerifiedProjection(changed.hash));
-    }
-    let projected_changed = graph
-        .view_header_node(changed.hash)
-        .ok_or(InvariantViolation::Protected(changed.hash))?;
-    if projected_changed != changed {
-        return Err(InvariantViolation::Index(changed.hash));
-    }
-
-    verify_indexes(engine_before_commit, plan)?;
-    let selected = engine_before_commit.selected_projection();
-    for hash in &plan.change_set.delete_nodes {
-        let Some(node) = engine_before_commit.graph().header_node(*hash) else {
-            return Err(InvariantViolation::Index(*hash));
-        };
-        if node.height >= finalized.height
-            && selected
-                .binary_search_by_key(&node.height, |frontier| frontier.height)
-                .ok()
-                .is_some_and(|index| selected[index].hash == *hash)
-        {
-            return Err(InvariantViolation::SelectedProjection(*hash));
-        }
-    }
-
-    let best = graph
-        .view_select_best_header_chain()
-        .map_err(|_| InvariantViolation::Selection)?;
-    if best.0 != metadata.frontiers.header_best || best.1 != metadata.header_best_score {
-        return Err(InvariantViolation::Selection);
-    }
-    if let Ok(index) = plan
-        .trust_pins
-        .binary_search_by_key(&finalized.height, |pin| pin.height)
-    {
-        if plan.trust_pins[index].hash != finalized.hash {
-            return Err(InvariantViolation::TrustPin(finalized.height));
-        }
-    }
-    verify_protected(graph, plan)?;
-    if graph.view_header_node_count().saturating_sub(1) > plan.limits.max_non_finalized_nodes.get()
-        || graph.view_eligible_header_tips().len() > plan.limits.max_candidate_tips.get()
-    {
-        return Err(InvariantViolation::Limits);
-    }
-    verify_generations(
-        engine_before_commit,
-        plan,
-        &[metadata.frontiers.header_best],
-        &[metadata.frontiers.verified_best],
-    )?;
-    verify_aux(engine_before_commit, graph, plan, mode)?;
-    Ok(())
-}
-
-fn verify_node<G: HeaderGraphView>(
-    graph: &G,
-    node: &HeaderNode,
-    work_origin: block::Hash,
-) -> Result<(), InvariantViolation> {
-    if node.header.hash() != node.hash {
-        return Err(InvariantViolation::NodeHash(node.hash));
-    }
-    if !graph
-        .view_header_hashes_at_height(node.height)
-        .contains(&node.hash)
-    {
-        return Err(InvariantViolation::Index(node.hash));
-    }
-    if node.work_coordinate().origin_hash() != work_origin {
-        return Err(InvariantViolation::Work(node.hash));
-    }
-    if node.hash == graph.view_finalized_frontier().hash {
-        if node.eligibility.inherited_from.is_some() {
-            return Err(InvariantViolation::Eligibility(node.hash));
-        }
-        return Ok(());
-    }
-    let parent = graph
-        .view_header_node(node.parent_hash)
-        .ok_or(InvariantViolation::Parent(node.hash))?;
-    if parent.height.next().ok() != Some(node.height)
-        || !graph.view_header_children(parent.hash).contains(&node.hash)
-    {
-        return Err(InvariantViolation::Parent(node.hash));
-    }
-    if parent.work_coordinate().checked_add(node.block_work).ok() != Some(node.work_coordinate()) {
-        return Err(InvariantViolation::Work(node.hash));
-    }
-    if node.eligibility.inherited_from != (!parent.is_eligible()).then_some(parent.hash) {
-        return Err(InvariantViolation::Eligibility(node.hash));
-    }
-    Ok(())
-}
-
-fn verify_indexes(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-) -> Result<(), InvariantViolation> {
-    if plan.change_set.put_nodes != plan.graph_delta().updated_header_nodes()
-        || plan.change_set.delete_nodes != plan.graph_delta().deleted_header_hashes()
-        || plan.change_set.put_consensus_invalid_body_tombstones
-            != plan.graph_delta().new_consensus_invalid_body_tombstones()
-    {
-        return Err(InvariantViolation::Index(block::Hash([0; 32])));
-    }
-    let mut inserted = HashSet::new();
-    for node in &plan.change_set.put_nodes {
-        if engine_before_commit.graph().header_node(node.hash).is_none() {
-            inserted.insert(Frontier::new(node.height, node.hash));
-        }
-    }
-    let indexed: HashSet<_> = plan
-        .change_set
-        .index_changes
-        .inserted
-        .iter()
-        .copied()
-        .collect();
-    if inserted != indexed {
-        return Err(InvariantViolation::Index(
-            inserted
-                .symmetric_difference(&indexed)
-                .next()
-                .map_or(block::Hash([0; 32]), |frontier| frontier.hash),
-        ));
-    }
-    let deleted: HashSet<_> = plan.change_set.delete_nodes.iter().copied().collect();
-    let deindexed: HashSet<_> = plan
-        .change_set
-        .index_changes
-        .deleted
-        .iter()
-        .copied()
-        .collect();
-    if deleted != deindexed {
-        return Err(InvariantViolation::Index(
-            deleted
-                .symmetric_difference(&deindexed)
-                .next()
-                .copied()
-                .unwrap_or(block::Hash([0; 32])),
-        ));
-    }
-    Ok(())
-}
-
-fn projected_path(
-    engine_before_commit: &HeaderChainEngine,
-    source: &crate::EngineSnapshot,
-    delta: &ProjectionDelta,
-    selected: bool,
-) -> Result<Vec<Frontier>, InvariantViolation> {
-    let tip = if selected {
-        source.frontiers.header_best
-    } else {
-        source.frontiers.verified_best
-    };
-    let mut path = if selected {
-        engine_before_commit.selected_projection().to_vec()
-    } else {
-        engine_before_commit.verified_projection().to_vec()
-    };
-    if path.last().copied() != Some(tip)
-        || path.first().copied() != Some(source.frontiers.finalized)
-    {
-        return Err(InvariantViolation::SnapshotBeforeCommit);
-    }
-    if let Some(remove_before) = delta.remove_before {
-        path.retain(|frontier| frontier.height >= remove_before);
-    }
-    if let Some(remove_from) = delta.remove_from {
-        path.retain(|frontier| frontier.height < remove_from);
-    }
-    path.extend(delta.put.iter().copied());
-    Ok(path)
-}
-
-fn verify_projection<G: HeaderGraphView>(
-    graph: &G,
-    projection: &[Frontier],
-    tip: Frontier,
-    failure: fn(block::Hash) -> InvariantViolation,
-) -> Result<(), InvariantViolation> {
-    if projection.first().copied() != Some(graph.view_finalized_frontier())
-        || projection.last().copied() != Some(tip)
-    {
-        return Err(failure(tip.hash));
-    }
-    for pair in projection.windows(2) {
-        if pair[1].height.0 != pair[0].height.0 + 1
-            || graph
-                .view_header_node(pair[1].hash)
-                .is_none_or(|node| node.parent_hash != pair[0].hash)
-        {
-            return Err(failure(pair[1].hash));
-        }
-    }
-    Ok(())
-}
-
-fn verify_verified<G: HeaderGraphView>(
-    graph: &G,
-    mode: EngineMode,
-    projection: &[Frontier],
-    tip: Frontier,
-) -> Result<(), InvariantViolation> {
-    verify_projection(
-        graph,
-        projection,
-        tip,
-        InvariantViolation::VerifiedProjection,
-    )?;
-    if mode == EngineMode::HeadersOnly && projection != [graph.view_finalized_frontier()] {
-        return Err(InvariantViolation::VerifiedProjection(tip.hash));
-    }
-    if mode == EngineMode::Integrated {
-        for frontier in projection.iter().skip(1) {
-            if !matches!(
-                graph
-                    .view_header_node(frontier.hash)
-                    .map(|node| node.body_validation_state.clone()),
-                Some(BodyValidationState::Verified { .. })
-            ) {
-                return Err(InvariantViolation::VerifiedProjection(frontier.hash));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn verify_pins(
-    pins: &[Frontier],
-    selected: &[Frontier],
-    verified: &[Frontier],
-    changed_nodes: &[HeaderNode],
-) -> Result<(), InvariantViolation> {
-    for pin in pins {
-        for projection in [selected, verified] {
-            if let Ok(index) =
-                projection.binary_search_by_key(&pin.height, |frontier| frontier.height)
-            {
-                let frontier = projection[index];
-                if frontier.hash != pin.hash {
-                    return Err(InvariantViolation::TrustPin(pin.height));
-                }
-            }
-        }
-        for node in changed_nodes
-            .iter()
-            .filter(|node| node.height == pin.height && node.hash != pin.hash)
-        {
-            let has_reason = node.eligibility.direct_reasons.iter().any(|reason| {
-                matches!(reason,
-                    EligibilityReason::SettledUpgradeConflict { height, expected }
-                    | EligibilityReason::CheckpointConflict { height, expected }
-                    if *height == pin.height && *expected == pin.hash)
-            });
-            if !has_reason {
-                return Err(InvariantViolation::TrustPin(pin.height));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn verify_protected<G: HeaderGraphView>(
-    graph: &G,
-    plan: &PlanCandidate,
-) -> Result<(), InvariantViolation> {
-    for frontier in [
-        plan.change_set.metadata.frontiers.finalized,
-        plan.change_set.metadata.frontiers.header_best,
-        plan.change_set.metadata.frontiers.verified_best,
-    ] {
-        if graph.view_header_node(frontier.hash).is_none()
-            || plan.change_set.delete_nodes.contains(&frontier.hash)
-        {
-            return Err(InvariantViolation::Protected(frontier.hash));
-        }
-    }
-    Ok(())
-}
-
-fn verify_generations(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-    selected: &[Frontier],
-    verified: &[Frontier],
-) -> Result<(), InvariantViolation> {
-    let old_selected = plan.snapshot_before_commit.frontiers.header_best;
-    let old_verified = plan.snapshot_before_commit.frontiers.verified_best;
-    let selected_changed = selected.last().copied() != Some(old_selected)
-        || !plan.change_set.selected_projection.put.is_empty()
-        || plan.change_set.selected_projection.remove_before.is_some()
-        || plan.change_set.selected_projection.remove_from.is_some();
-    let verified_changed = verified.last().copied() != Some(old_verified)
-        || !plan.change_set.verified_projection.put.is_empty()
-        || plan.change_set.verified_projection.remove_before.is_some()
-        || plan.change_set.verified_projection.remove_from.is_some();
-    let alarm_changed = plan.snapshot_before_commit.alarms != plan.change_set.metadata.alarms;
-    let effects = !plan.change_set.put_nodes.is_empty()
-        || !plan.change_set.delete_nodes.is_empty()
-        || !plan.change_set.aux_changes.is_empty()
-        || plan.change_set.finality_append.is_some()
-        || selected_changed
-        || verified_changed
-        || alarm_changed;
-    let expected_state = if effects {
-        plan.snapshot_before_commit.state_version.checked_next().ok()
-    } else {
-        Some(plan.snapshot_before_commit.state_version)
-    };
-    let header_validation_changed =
-        plan.change_set
-            .put_nodes
-            .iter()
-            .try_fold(false, |changed, node| {
-                Ok::<_, InvariantViolation>(
-                    changed
-                        || engine_before_commit
-                            .graph()
-                            .header_node(node.hash)
-                            .is_some_and(|old| old.validation != node.validation),
-                )
-            })?;
-    let header_eligibility_changed = plan.change_set.put_nodes.iter().any(|node| {
-        engine_before_commit
-            .graph()
-            .header_node(node.hash)
-            .is_some_and(|old| old.is_eligible() != node.is_eligible())
-    });
-    let header_effect = selected_changed
-        || !plan.change_set.index_changes.inserted.is_empty()
-        || !plan.change_set.delete_nodes.is_empty()
-        || header_validation_changed
-        || header_eligibility_changed
-        || !plan.change_set.eligibility_changes.is_empty()
-        || plan.change_set.finality_append.is_some();
-    let expected_header = if header_effect {
-        plan.snapshot_before_commit.header_generation.checked_next().ok()
-    } else {
-        Some(plan.snapshot_before_commit.header_generation)
-    };
-    let verified_effect = verified_changed || plan.change_set.finality_append.is_some();
-    let expected_verified = if verified_effect {
-        plan.snapshot_before_commit.verified_generation.checked_next().ok()
-    } else {
-        Some(plan.snapshot_before_commit.verified_generation)
-    };
-    if Some(plan.change_set.metadata.state_version) != expected_state
-        || Some(plan.change_set.metadata.header_generation) != expected_header
-        || Some(plan.change_set.metadata.verified_generation) != expected_verified
-    {
-        return Err(InvariantViolation::Generation);
-    }
-    Ok(())
-}
-
-fn verify_aux<G: HeaderGraphView>(
-    engine_before_commit: &HeaderChainEngine,
-    graph: &G,
-    plan: &PlanCandidate,
-    mode: VerificationMode,
-) -> Result<(), InvariantViolation> {
-    let mut put_ids = HashSet::new();
-    for change in &plan.change_set.aux_changes {
-        let AuxDelta::Put(delivery) = change else {
-            continue;
-        };
-        if !put_ids.insert(delivery.delivery_id)
-            || engine_before_commit
-                .aux_delivery(delivery.delivery_id)
-                .is_some_and(|existing| existing.header_hash != delivery.header_hash)
-        {
-            return Err(InvariantViolation::Auxiliary(delivery.header_hash));
-        }
-    }
-    let deletes: Vec<_> = plan
-        .change_set
-        .aux_changes
-        .iter()
-        .filter_map(|change| match change {
-            AuxDelta::Delete {
-                header_hash,
-                delivery_id,
-            } => Some((*header_hash, *delivery_id)),
-            AuxDelta::Put(_) => None,
-        })
-        .collect();
-    for (header_hash, delivery_id) in &deletes {
-        let exists = engine_before_commit
-            .aux_deliveries(*header_hash)
-            .iter()
-            .any(|delivery| delivery.delivery_id == *delivery_id);
-        if !exists {
-            return Err(InvariantViolation::Auxiliary(*header_hash));
-        }
-    }
-    let deleted_ids: HashSet<_> = deletes
-        .into_iter()
-        .map(|(_, delivery_id)| delivery_id)
-        .collect();
-    let puts: HashMap<_, _> = plan
-        .change_set
-        .aux_changes
-        .iter()
-        .filter_map(|change| match change {
-            AuxDelta::Put(delivery) => Some((delivery.delivery_id, delivery.as_ref())),
-            AuxDelta::Delete { .. } => None,
-        })
-        .collect();
-    let nodes: Vec<&HeaderNode> = match mode {
-        #[cfg(any(test, feature = "fuzz-impl"))]
-        VerificationMode::Exhaustive => graph.view_header_nodes(),
-        #[cfg(any(test, not(feature = "fuzz-impl")))]
-        VerificationMode::Production => plan
-            .change_set
-            .put_nodes
-            .iter()
-            .filter_map(|changed| graph.view_header_node(changed.hash))
-            .collect(),
-    };
-    for node in nodes {
-        let mut deliveries = engine_before_commit.aux_deliveries(node.hash).to_vec();
-        deliveries.retain(|delivery| !deleted_ids.contains(&delivery.delivery_id));
-        deliveries.extend(
-            puts.values()
-                .filter(|delivery| delivery.header_hash == node.hash)
-                .map(|delivery| **delivery),
-        );
-        for delivery in deliveries {
-            if delivery.header_hash != node.hash
-                || !node.aux_delivery_ids.contains(&delivery.delivery_id)
-            {
-                return Err(InvariantViolation::Auxiliary(node.hash));
-            }
-        }
-    }
-    for delivery in puts.values() {
-        if graph.view_header_node(delivery.header_hash).is_none() {
-            return Err(InvariantViolation::Auxiliary(delivery.header_hash));
-        }
-    }
-    for hash in &plan.change_set.delete_nodes {
-        for delivery in engine_before_commit.aux_deliveries(*hash) {
-            if !deleted_ids.contains(&delivery.delivery_id) {
-                return Err(InvariantViolation::Auxiliary(*hash));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn verification_nodes<'a, G: HeaderGraphView>(
-    _engine_before_commit: &HeaderChainEngine,
-    graph: &'a G,
-    _plan: &PlanCandidate,
-    mode: VerificationMode,
-) -> Vec<&'a HeaderNode> {
-    match mode {
-        #[cfg(any(test, feature = "fuzz-impl"))]
-        VerificationMode::Exhaustive => graph.view_header_nodes(),
-        #[cfg(any(test, not(feature = "fuzz-impl")))]
-        VerificationMode::Production => changed_boundary_nodes(_engine_before_commit, graph, _plan),
-    }
-}
-
-#[cfg(any(test, feature = "fuzz-impl"))]
-fn materialize_projected_graph(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &PlanCandidate,
-) -> Result<crate::graph::MemHeaderStore, InvariantViolation> {
-    let mut graph = engine_before_commit.graph().clone();
-    graph
-        .apply_delta(plan.graph_delta())
-        .map_err(|_| InvariantViolation::Index(block::Hash([0; 32])))?;
-    Ok(graph)
-}
-
-#[cfg(any(test, not(feature = "fuzz-impl")))]
-fn changed_boundary_nodes<'a, G: HeaderGraphView>(
-    engine_before_commit: &HeaderChainEngine,
-    graph: &'a G,
-    plan: &PlanCandidate,
-) -> Vec<&'a HeaderNode> {
-    let mut hashes = HashSet::from([
-        plan.change_set.metadata.frontiers.finalized.hash,
-        plan.change_set.metadata.frontiers.header_best.hash,
-        plan.change_set.metadata.frontiers.verified_best.hash,
-    ]);
-    for node in plan.graph_delta().updated_header_nodes() {
-        hashes.insert(node.hash);
-        hashes.insert(node.parent_hash);
-        hashes.extend(graph.view_header_children(node.hash));
-    }
-    for hash in plan.graph_delta().deleted_header_hashes() {
-        if let Some(node) = engine_before_commit.graph().header_node(*hash) {
-            hashes.insert(node.parent_hash);
-            hashes.extend(engine_before_commit.graph().header_children(*hash));
-        }
-    }
-    hashes
-        .into_iter()
-        .filter_map(|hash| graph.view_header_node(hash))
-        .collect()
-}
-
-/// Re-run verification against an already verified plan in tests and fuzzing.
-#[cfg(test)]
-pub(crate) fn verify_plan(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &TransitionPlan,
-) -> Result<(), InvariantViolation> {
-    verify_plan_with_mode(engine_before_commit, plan.candidate(), default_verification_mode())
-}
-
-/// Verify `plan` using the exact production overlay and boundary-node path.
-#[cfg(test)]
-pub(crate) fn verify_plan_production(
-    engine_before_commit: &HeaderChainEngine,
-    plan: &TransitionPlan,
-) -> Result<(), InvariantViolation> {
-    verify_plan_with_mode(engine_before_commit, plan.candidate(), VerificationMode::Production)
-}
