+++ crates/zakura-header-chain/src/transition/invariants/checks/mod.rs
@@ -1,0 +1,17 @@
+//! Leaf commit-invariant checks grouped by responsibility.
+
+mod aux;
+mod generations;
+mod indexes;
+mod nodes;
+mod pins;
+mod projections;
+mod protected;
+
+pub(crate) use aux::verify_aux;
+pub(crate) use generations::verify_generations;
+pub(crate) use indexes::verify_indexes;
+pub(crate) use nodes::verify_node;
+pub(crate) use pins::verify_pins;
+pub(crate) use projections::{projected_path, verify_projection, verify_verified};
+pub(crate) use protected::verify_protected;
